requirements.txt
psutil>=5.9.4
pyte>=0.8.1
PyQt6>=6.5.2 
PyQt6.QScintilla>=2.14.0
hy>=0.25.0
cx_freeze>=6.15.0
fpdf2>=2.7.8
black
autopep8
yapf
clang-format
beautifulsoup4
ruff

# Linux
ptyprocess>=0.7.0; platform_system == "Linux"

# Windows
pywin32>=306; platform_system == "Windows"
pywinpty>=2.0.11; platform_system == "Windows"
